requirements.txt
Flask==2.3.2
Flask-Caching==2.0.2
requests==2.31.0
pycryptodome==3.19.0
colorama==0.4.6
protobuf==4.24.4
gunicorn==21.2.0
cachetools
urllib3
aiohttp
PyGithub==2.3.0

